requirements.txt
pikepdf
pymupdf
Pillow
